Refactor: influxDB version 변경으로 인한 Temperature api 수정

+++ src/main/java/com/nhnacademy/sensordata/entity/temperature/Temperature.java
@@ -23,5 +23,5 @@ public class Temperature {
     @Column(name = "topic")
     private String topic;
     @Column(name = "value")
-    private Double value;
+    private Float value;
 }
